New internalConsistent member of SafeCopy.
Ignore-this: 4cd6fe99f39c9d72f3032068fccb12cc

As it is now, 'checkInvariants' always have to construct the list
of versions and check it, which can be costly if safePut/safeGet
is used many times.  Even if in practice the '[a]' instance uses
getSafeGet/getSafePut which incur the cost of 'checkInvariants'
only once per list, this should be fixed.

The internalConsistent member of SafeCopy basically puts the
information of weather everything is fine inside the type class.
This means that the information only has to be calculated only
for the whole program (since it is a single value, not a
function).  Then 'checkInvariant' only needs to verify a single
boolean value.

darcs-hash:20110418002043-0b5ec-384ef6d98fbb311c2568e13957d6b095cb99e5bd