Use getSafeGet and getSafePut on derived instances.
Ignore-this: 7a9df79e0e0ba4d2c9b8933c3895990c

For example, take the data type

  data T a b = C a Int a

Instead of generating

  putCopy (C x y z) = do safePut x
                         safePut y
                         safePut z

we generate

  putCopy (C x y z) = do safePut_a   <- getSafePut
                         safePut_Int <- getSafePut
                         safePut_a   x
		         safePut_Int y
                         safePut_a   z

which in this case saves four bytes.  Note that we only get
functions for the types that are needed; in this case, we don't
get a safePut for the type 'b', as it isn't used.

The same happens for getCopy.  Instead of

  getCopy = return C <*> safeGet <*> safeGet <*> safeGet

we generate

  getCopy = do safeGet_a   <- getSafeGet
               safeGet_Int <- getSafeGet
               return C <*> safeGet_a <*> safeGet_Int <*> safeGet_a

darcs-hash:20110418131226-0b5ec-57f547a7e11a67e8b014f60039493ffc53a665a9